+++ test/validator/query.test.ts
@@ -1,4 +1,4 @@
-import { Elysia, t } from '../../src'
+import { Context, Elysia, t } from '../../src'
 
 import { describe, expect, it } from 'bun:test'
 import { req } from '../utils'
@@ -234,5 +234,38 @@ describe('Query Validator', () => {
 			name: 'nagisa',
 			rank: 1
 		})
+	})
+
+	it('handle query edge case', async () => {
+		const checker = {
+			check(ctx: Context, name: string, state?: string) {
+				return typeof state !== 'undefined'
+			}
+		}
+
+		const app = new Elysia()
+			.derive((ctx) => {
+				const { name } = ctx.params
+
+				return {
+					check() {
+						const { state } = ctx.query
+
+						if (!checker.check(ctx, name, state ?? ctx.query.state))
+							throw new Error('State mismatch')
+					}
+				}
+			})
+			.get('/:name', ({ check }) => {
+				check()
+
+				return 'yay'
+			})
+
+		const response = await app
+			.handle(req('/a?state=123'))
+			.then((x) => x.text())
+
+		expect(response).toBe('yay')
 	})
 })